refactor repairDuplicateVersions to avoid main() running when importing its functions
