+++ tests/conf/envSetup.js
@@ -1,0 +1,2 @@
+process.env.BUCKETD_HOSTPORT = 'localhost:9000';
+process.env.SPROXYD_HOSTPORT = 'localhost';
